test: start fixing the test suite again

+++ lib/orocos/test.rb
@@ -4,6 +4,17 @@ require 'orogen'
 
 module Orocos
     module Test
+        USE_MQUEUE =
+            if ENV['USE_MQUEUE'] == '1'
+                puts "MQueue enabled through the USE_MQUEUE environment variable"
+                puts "set USE_MQUEUE=0 to disable"
+                true
+            else
+                puts "use of MQueue disabled. Set USE_MQUEUE=1 to enable"
+                false
+            end
+
+
         attr_reader :processes
 
         def setup
@@ -29,14 +40,21 @@ module Orocos
 
             FileUtils.mkdir_p work_basedir
             work_dir = File.join(work_basedir, src_name)
-            FileUtils.rm_rf work_dir
-            FileUtils.cp_r  src_dir, work_dir
+            if (ENV['TEST_KEEP_WC'] != "1") || !File.directory?(work_dir)
+                FileUtils.rm_rf work_dir
+                FileUtils.cp_r  src_dir, work_dir
+            end
 
             prefix   = File.join(work_basedir, "prefix")
             ruby_bin   = RbConfig::CONFIG['RUBY_INSTALL_NAME']
             orogen_bin = File.expand_path('../bin/orogen', Orocos::Generation.base_dir)
             Dir.chdir(work_dir) do
-                if !system(ruby_bin, orogen_bin, '--corba', '--transports=corba,typelib,mqueue', File.basename(src))
+                transports = %w{corba typelib}
+                if Test::USE_MQUEUE
+                    transports << mqueue
+                end
+
+                if !system(ruby_bin, orogen_bin, '--corba', "--transports=#{transports.join(",")}", File.basename(src))
                     raise "failed to build #{src} in #{work_basedir}"
                 end
 
@@ -70,20 +88,9 @@ module Orocos
     end
 
     module Spec
-        USE_MQUEUE =
-            if ENV['USE_MQUEUE'] == '1'
-                puts "MQueue enabled through the USE_MQUEUE environment variable"
-                puts "set USE_MQUEUE=0 to disable"
-                true
-            else
-                puts "use of MQueue disabled. Set USE_MQUEUE=1 to enable"
-                false
-            end
-
-
         def setup
             ENV['PKG_CONFIG_PATH'] = "#{File.join(WORK_DIR, "prefix", 'lib', 'pkgconfig')}"
-            Orocos::MQueue.auto = USE_MQUEUE
+            Orocos::MQueue.auto = Test::USE_MQUEUE
             Orocos.initialize
             Orocos.export_types = false
